refactor(wingman): extract WingmanMcpManager service
Move MCP discovery, enable/disable, and connection orchestration out
of the unified Wingman class. The manager now owns the McpRegistry
and the secret/header/timeout handling for connecting to servers.
