+++ wingmen/wingman.py
@@ -41,7 +41,6 @@ from api.enums import (
     TtsProvider,
     WingmanInitializationErrorType,
 )
-from api.commands import McpStateChangedCommand
 from providers.interfaces import LlmInterface, SttInterface, TtsInterface
 from services.audio_player import AudioPlayer
 from services.benchmark import Benchmark
@@ -58,8 +57,8 @@ from services.tool_executor import ToolExecutor
 from services.provider_factory import ProviderFactory
 from services.skill_registry import SkillRegistry
 from services.mcp_client import McpClient
-from services.mcp_registry import McpRegistry
 from services.capability_registry import CapabilityRegistry
+from services.wingman_mcp_manager import WingmanMcpManager
 from services.tool_response_cache import ToolResponseCompressor
 from services.token_utils import count_tokens
 from skills.skill_base import Skill
@@ -168,10 +167,13 @@ class Wingman:
 
         # --- MCP ---
         self.mcp_client = McpClient(wingman_name=self.name)
-        self.mcp_registry = McpRegistry(
-            self.mcp_client,
+        self.mcp_manager = WingmanMcpManager(
             wingman_name=self.name,
-            on_state_changed=self._broadcast_mcp_state_changed,
+            mcp_client=self.mcp_client,
+            secret_keeper=self.secret_keeper,
+            get_mcp_config=lambda: self.tower.config_manager.mcp_config if self.tower else None,
+            settings=self.settings,
+            config=self.config,
         )
 
         # --- Unified capability registry ---
@@ -190,6 +192,13 @@ class Wingman:
         self.last_gpt_call = None
         self.instant_responses = []
         self.last_used_instant_responses = []
+
+    # ──────────────────────────────── Backward-compat properties ──────────────── #
+
+    @property
+    def mcp_registry(self):
+        """Backward-compat: many callers access wingman.mcp_registry directly."""
+        return self.mcp_manager.mcp_registry
 
     # ──────────────────────────────── Record keys ─────────────────────────────── #
 
@@ -343,7 +352,7 @@ class Wingman:
         self.skill_registry.clear()
 
     async def unload_mcps(self):
-        await self.mcp_registry.clear()
+        return await self.mcp_manager.unload_mcps()
 
     # ──────────────────────────────── Memory ──────────────────────────────────── #
 
@@ -365,236 +374,16 @@ class Wingman:
             return True
         return False
 
-    # ──────────────────────────────── MCP ──────────────────────────────────────── #
-
-    def _broadcast_mcp_state_changed(self):
-        if printr._connection_manager:
-            printr.ensure_async(
-                printr._connection_manager.broadcast(
-                    McpStateChangedCommand(wingman_name=self.name)
-                )
-            )
+    # ──────────────────────────────── MCP (forwarding) ─────────────────────────── #
 
     async def enable_mcp(self, mcp_name: str) -> tuple[bool, str]:
-        if not self.mcp_client.is_available:
-            return False, "MCP SDK not installed."
-
-        if mcp_name in self.mcp_registry.get_connected_server_names():
-            return True, f"MCP server '{mcp_name}' is already connected."
-
-        central_mcp_config = self.tower.config_manager.mcp_config
-        mcp_configs = central_mcp_config.servers if central_mcp_config else []
-
-        mcp_config = None
-        for cfg in mcp_configs:
-            if cfg.name == mcp_name:
-                mcp_config = cfg
-                break
-
-        if not mcp_config:
-            return False, f"MCP server '{mcp_name}' not found in mcp.yaml."
-
-        try:
-            headers = {}
-            if mcp_config.headers:
-                headers.update(mcp_config.headers)
-
-            secret_key = f"mcp_{mcp_config.name}"
-            api_key = await self.secret_keeper.retrieve(
-                requester=self.name,
-                key=secret_key,
-                prompt_if_missing=False,
-            )
-            if api_key:
-                if not any(
-                    k.lower() in ["authorization", "api-key", "x-api-key"]
-                    for k in headers.keys()
-                ):
-                    headers["Authorization"] = f"Bearer {api_key}"
-
-            default_timeout = 60.0 if mcp_config.type.value == "stdio" else 30.0
-            timeout = (
-                float(mcp_config.timeout) if mcp_config.timeout else default_timeout
-            )
-
-            connection = await asyncio.wait_for(
-                self.mcp_registry.register_server(
-                    config=mcp_config,
-                    headers=headers if headers else None,
-                ),
-                timeout=timeout,
-            )
-
-            if connection.is_connected:
-                tool_count = len(connection.tools)
-                return True, f"MCP server '{mcp_name}' enabled with {tool_count} tools."
-            else:
-                error = connection.error or "Connection failed."
-                return False, f"MCP server '{mcp_name}' failed to connect: {error}"
-
-        except asyncio.TimeoutError:
-            error_msg = f"Connection timed out ({int(timeout)}s)."
-            self.mcp_registry.set_server_error(mcp_name, error_msg)
-            return False, f"MCP server '{mcp_name}': {error_msg}"
-
-        except Exception as e:
-            error_msg = f"Error enabling MCP '{mcp_name}': {str(e)}"
-            await printr.print_async(error_msg, color=LogType.ERROR)
-            printr.print(traceback.format_exc(), color=LogType.ERROR, server_only=True)
-            return False, error_msg
+        return await self.mcp_manager.enable_mcp(mcp_name)
 
     async def disable_mcp(self, mcp_name: str) -> tuple[bool, str]:
-        if mcp_name not in self.mcp_registry.get_connected_server_names():
-            return True, f"MCP server '{mcp_name}' is already disconnected."
-
-        try:
-            await self.mcp_registry.unregister_server(mcp_name)
-            return True, f"MCP server '{mcp_name}' disabled."
-
-        except Exception as e:
-            error_msg = f"Error disabling MCP '{mcp_name}': {str(e)}"
-            await printr.print_async(error_msg, color=LogType.ERROR)
-            printr.print(traceback.format_exc(), color=LogType.ERROR, server_only=True)
-            return False, error_msg
+        return await self.mcp_manager.disable_mcp(mcp_name)
 
     async def init_mcps(self) -> list[WingmanInitializationError]:
-        errors = []
-
-        if not self.mcp_client.is_available:
-            printr.print(
-                f"[{self.name}] MCP SDK not installed, skipping MCP initialization.",
-                color=LogType.WARNING,
-                server_only=True,
-            )
-            return errors
-
-        await self.unload_mcps()
-
-        central_mcp_config = self.tower.config_manager.mcp_config
-        mcp_configs = central_mcp_config.servers if central_mcp_config else []
-        if not mcp_configs:
-            return errors
-
-        discoverable_mcps = self.config.discoverable_mcps
-        mcps_to_connect = [mcp for mcp in mcp_configs if mcp.name in discoverable_mcps]
-
-        if not mcps_to_connect:
-            return errors
-
-        async def connect_mcp(mcp_config):
-            local_errors = []
-            try:
-                headers = {}
-                if mcp_config.headers:
-                    headers.update(mcp_config.headers)
-
-                secret_key = f"mcp_{mcp_config.name}"
-                api_key = await self.secret_keeper.retrieve(
-                    requester=self.name,
-                    key=secret_key,
-                    prompt_if_missing=False,
-                )
-                if api_key:
-                    printr.print(
-                        f"MCP secret '{secret_key}' found ({len(api_key)} chars)",
-                        color=LogType.INFO,
-                        source_name=self.name,
-                        server_only=True,
-                    )
-                    if not any(
-                        k.lower() in ["authorization", "api-key", "x-api-key"]
-                        for k in headers.keys()
-                    ):
-                        headers["Authorization"] = f"Bearer {api_key}"
-
-                default_timeout = 60.0 if mcp_config.type.value == "stdio" else 30.0
-                timeout = (
-                    float(mcp_config.timeout) if mcp_config.timeout else default_timeout
-                )
-
-                try:
-                    connection = await asyncio.wait_for(
-                        self.mcp_registry.register_server(
-                            config=mcp_config,
-                            headers=headers if headers else None,
-                        ),
-                        timeout=timeout,
-                    )
-                except asyncio.TimeoutError:
-                    error_msg = f"MCP '{mcp_config.display_name}' connection timed out ({int(timeout)}s)."
-                    printr.print(
-                        error_msg,
-                        color=LogType.WARNING,
-                        source_name=self.name,
-                        server_only=True,
-                    )
-                    local_errors.append(
-                        WingmanInitializationError(
-                            wingman_name=self.name,
-                            message=error_msg,
-                            error_type=WingmanInitializationErrorType.MCP_CONNECTION_FAILED,
-                        )
-                    )
-                    return (False, None, local_errors)
-
-                if connection.is_connected:
-                    return (
-                        True,
-                        f"{mcp_config.display_name} ({len(connection.tools)} tools)",
-                        local_errors,
-                    )
-                else:
-                    error_msg = f"MCP '{mcp_config.display_name}' failed to connect: {connection.error}"
-                    local_errors.append(
-                        WingmanInitializationError(
-                            wingman_name=self.name,
-                            message=error_msg,
-                            error_type=WingmanInitializationErrorType.MCP_CONNECTION_FAILED,
-                        )
-                    )
-                    return (False, None, local_errors)
-
-            except Exception as e:
-                error_msg = f"MCP '{mcp_config.name}' initialization error: {str(e)}"
-                printr.print(
-                    error_msg,
-                    color=LogType.ERROR,
-                    source_name=self.name,
-                    server_only=True,
-                )
-                printr.print(
-                    traceback.format_exc(), color=LogType.ERROR, server_only=True
-                )
-                local_errors.append(
-                    WingmanInitializationError(
-                        wingman_name=self.name,
-                        message=error_msg,
-                        error_type=WingmanInitializationErrorType.MCP_CONNECTION_FAILED,
-                    )
-                )
-                return (False, None, local_errors)
-
-        connection_tasks = [connect_mcp(mcp) for mcp in mcps_to_connect]
-        results = await asyncio.gather(*connection_tasks)
-
-        connected_count = 0
-        connected_names = []
-        for success, connection_info, mcp_errors in results:
-            if success:
-                connected_count += 1
-                connected_names.append(connection_info)
-            errors.extend(mcp_errors)
-
-        if connected_count > 0:
-            await printr.print_async(
-                f"Discoverable MCP servers connected ({connected_count}): {', '.join(connected_names)}",
-                color=LogType.WINGMAN,
-                source=LogSource.WINGMAN,
-                source_name=self.name,
-                server_only=not self.settings.debug_mode,
-            )
-
-        return errors
+        return await self.mcp_manager.init_mcps()
 
     # ──────────────────────────────── Skills ───────────────────────────────────── #
